Specification Review
This contains all of the specifications that i looked over. I
highlighted the ones i wanted to ask the PM before he/she left for the
airpoint. I created comments within the word document that explain each
of my questions/concerns with the specifications that i asked the PM
about.